@@ -3,7 +3,7 @@ description: example
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
   bloc_analytics:
