@@ -1,16 +1,16 @@
 name: bloc_analytics
 description: Easy integrate you app with multiple analytics tracker, integrate with Firebase Analytics, Crashlytics and others
-version: 0.2.0
+version: 0.3.0
 author: guitcastro <guilherme.torres.castro@gmail.com>
 homepage: https://github.com/guitcastro/flutter_bloc_analytics
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
-  meta: ^1.1.6
+  meta: ^1.1.8
 
 dev_dependencies:
-  test: ">=1.3.0 <2.0.0"
-  test_coverage: ^0.2.0
-  mockito: ^4.0.0
+  test: ">=1.10.0 <2.0.0"
+  test_coverage: ^0.4.0
+  mockito: ^4.1.1
